xyz/scansuite/preprocess-service/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
numpy
pillow
python-multipart
